1. optimize task cancel logic. 2. add wifi pwd wrong logic. 3. optimize task callback.

+++ esaywifilib/src/main/java/com/likang/easywifi/lib/core/task/ScanWifiTask.java
@@ -148,7 +148,7 @@ public class ScanWifiTask extends WifiTask {
         dest.writeLong(mScanWifiTimeout);
         dest.writeLong(mSetWifiEnabledTimeout);
         dest.writeInt(mScanWifiWay);
-        dest.writeInt((byte) (mIsAutoSwitchToThroughSystemWifi ? 1 : 0));
+        dest.writeByte((byte) (mIsAutoSwitchToThroughSystemWifi ? 1 : 0));
     }
 
     @Override
